Refactor everything into packages
