+++ src/main/java/Klalopz/Instructions/MarkInstruction.java
@@ -1,4 +1,10 @@
-import java.util.List;
+package Klalopz.Instructions;
+
+import Klalopz.Exceptions.KlalopzException;
+import Klalopz.Storage.DataStorage;
+import Klalopz.Tasks.Task;
+import Klalopz.Tasks.TaskList;
+import Klalopz.Ui.TextUi;
 
 public class MarkInstruction implements Instruction {
     public String arguments;
@@ -11,7 +17,7 @@ public class MarkInstruction implements Instruction {
         this.arguments = arguments;
         this.index = Integer.parseInt(arguments.trim()) - 1;
     }
-    public void execute(TaskList storage, DataStorage dataStorage, Ui ui) {
+    public void execute(TaskList storage, DataStorage dataStorage, TextUi ui) {
 
         Task currTask = storage.getTask(index);
         currTask.setCompleted(Boolean.TRUE);
