+++ src/main/java/Klalopz/Tasks/Deadline.java
@@ -1,3 +1,5 @@
+package Klalopz.Tasks;
+
 import java.time.LocalDate;
 
 public class Deadline extends Task {
