+++ src/main/java/Klalopz/Tasks/Event.java
@@ -1,3 +1,5 @@
+package Klalopz.Tasks;
+
 import java.time.LocalDate;
 
 public class Event extends Task {
