+++ src/main/java/Klalopz/Tasks/TaskList.java
@@ -1,3 +1,7 @@
+package Klalopz.Tasks;
+
+import Klalopz.Tasks.Task;
+
 import java.util.ArrayList;
 import java.util.List;
