+++ src/main/java/Klalopz/Tasks/ToDo.java
@@ -1,3 +1,5 @@
+package Klalopz.Tasks;
+
 public class ToDo extends Task {
 
     public ToDo(String description, Boolean isCompleted) {
